Adds the basics of a webrtc based interview tool
This is the outline for a recording tool that would allow us to record
interviews with interesting redecentralisation people without the need
to use Google Hangouts. Currently uses a STUN server hosted at
redecentralise.net

Recording in WebRTC still seems a bit up in the air, and so we need to
wait until someone implements MediaStream recording (see the  spec
at http://www.w3.org/TR/mediastream-recording/) so that we can have
the server join as a participant (and then save to disk).

For now, I've used recordrtc, that unfortunately only records with Chrome
which isn't great :( but does allow us to start/stop and save the video
but not the audio.  Audio has to be done separately so this isn't a good
solution for more than a proof of concept.

Will also need work to see if it is possible to switch between
video streams based on the when the host is not speaking. Without
that we'd just have to choose a single video-stream to record :(